+++ src/migrate.ts
@@ -1,91 +1,215 @@
 import { createRoutesFromFolders } from '@remix-run/v1-route-convention'
 import * as fs from 'fs'
 import * as path from 'path'
-import { getRouteSegments } from './index'
+import { type RouteManifest } from './index'
 import { defineRoutes } from './routes'
 
-export type RoutingConvention = 'flat-files' | 'flat-folders'
+export type RoutingConvention = 'flat-files' | 'flat-folders' | 'hybrid'
 export type MigrateOptions = {
   convention: RoutingConvention
+  force: boolean
+  ignoredRouteFiles?: string[]
 }
+
+const pathSepRegex = new RegExp(`\\${path.sep}`, 'g')
+const routeExtensions = ['.js', '.jsx', '.ts', '.tsx', '.md', '.mdx']
+
 export function migrate(
   sourceDir: string,
   targetDir: string,
-  options: MigrateOptions = { convention: 'flat-files' },
+  options: MigrateOptions = {
+    convention: 'flat-files',
+    force: false,
+    ignoredRouteFiles: undefined,
+  },
 ) {
-  // var visitor = options.convention === 'flat-files' ? flatFiles : flatFolders
+  if (sourceDir.startsWith('./')) {
+    sourceDir = sourceDir.substring(2)
+  }
+  if (targetDir.startsWith('./')) {
+    targetDir = targetDir.substring(2)
+  }
 
-  // visitFiles(sourceDir, visitor(sourceDir, targetDir))
+  if (!options.ignoredRouteFiles) {
+    // get remix.config.js
+    const remixConfigPath = path.join(process.cwd(), 'remix.config.js')
+    if (fs.existsSync(remixConfigPath)) {
+      const remixConfig = require(remixConfigPath)
+      options.ignoredRouteFiles = remixConfig.ignoredRouteFiles
+    }
+  }
+
+  console.log(
+    `🛠️ Migrating to flat-routes using ${options.convention} convention...`,
+  )
+  console.log(`🗂️ source: ${sourceDir}`)
+  console.log(`🗂️ target: ${targetDir}`)
+  console.log(`🙈ignored files: ${options.ignoredRouteFiles}`)
+  console.log()
 
   const routes = createRoutesFromFolders(defineRoutes, {
-    appDirectory: './app',
+    appDirectory: './',
     routesDirectory: sourceDir,
+    ignoredFilePatterns: options.ignoredRouteFiles,
   })
 
-  Object.entries(routes).forEach(([_key, route]) => {
-    let file = route.file
-    let flat = file
-      .replace(/[\/]/g, '.')
-      .replace(/\._([^_.])/g, '.[_]$1')
-      .replace(/\.__/g, '._')
-      .replace(/\.index.tsx$/, '._index.tsx')
-      .replace(/^routes\./, 'routes/')
-    console.log(flat)
-  })
-}
-
-export function flatFiles(sourceDir: string, targetDir: string) {
-  return (file: string) => {
-    console.log(file)
+  Object.entries(routes).forEach(([id, route]) => {
+    let { path: routePath, file, parentId } = route
     let extension = path.extname(file)
-    let name = file.substring(0, file.length - extension.length)
-    const route = convertToRoute(name)
-    const targetFile = path.join(targetDir, `${route}${extension}`)
-    fs.cpSync(path.join(sourceDir, file), targetFile, { force: true })
-  }
-}
-
-const routeExtensions = ['.js', '.jsx', '.ts', '.tsx', '.md', '.mdx']
-export function flatFolders(sourceDir: string, targetDir: string) {
-  return (file: string) => {
-    console.log(file)
-    const extension = path.extname(file)
-    const name = file.substring(0, file.length - extension.length)
-    const route = convertToRoute(name)
-    const targetFolder = path.join(targetDir, route)
+    // skip non-route files if not ignored above
     if (!routeExtensions.includes(extension)) {
       return
     }
-    fs.mkdirSync(targetFolder, { recursive: true })
-    fs.cpSync(
-      path.join(sourceDir, file),
-      path.join(targetFolder, `/_index${extension}`),
-      {
-        force: true,
-      },
+
+    let flat = convertToRoute(
+      routes,
+      sourceDir,
+      id,
+      parentId!,
+      routePath!,
+      !!route.index,
+      options.convention,
     )
+
+    // replace sourceDir with targetDir
+    flat = path.join(targetDir, flat)
+
+    switch (options.convention) {
+      case 'flat-folders': {
+        fs.mkdirSync(flat, { recursive: true })
+        fs.cpSync(file, path.join(flat, `/route${extension}`), {
+          force: true,
+        })
+        break
+      }
+      case 'hybrid': {
+        fs.mkdirSync(path.dirname(flat), { recursive: true })
+        const targetFile = `${flat}${extension}`
+        fs.cpSync(file, targetFile, { force: true })
+        break
+      }
+      case 'flat-files': {
+        const targetFile = `${flat}${extension}`
+        fs.cpSync(file, targetFile, { force: true })
+        break
+      }
+    }
+  })
+  console.log('🏁 Finished!')
+}
+
+export function convertToRoute(
+  routes: RouteManifest,
+  sourceDir: string,
+  id: string,
+  parentId: string,
+  routePath: string,
+  index: boolean,
+  convention: RoutingConvention,
+) {
+  // strip sourceDir from id and parentId
+  let routeId = id.substring(sourceDir.length + 1)
+  parentId =
+    parentId === 'root' ? parentId : parentId.substring(sourceDir.length + 1)
+
+  if (parentId && parentId !== 'root') {
+    if (convention !== 'hybrid' && routePath?.includes('/')) {
+      // multi-segment route, so need to fixup parent for flat-routes (trailing _)
+      // strip parent route from route
+      let currentPath = routeId.substring(parentId.length + 1)
+      let routeSegments = getRouteSegments(currentPath)
+      let dottedSegments = getEscapedDottedRouteSegments(routeSegments[0])
+      console.log({ routeSegments, dottedSegments })
+      if (dottedSegments.length > 1) {
+        const [first, ...rest] = dottedSegments
+        //rewrite id to use trailing _ for parent
+        routeId = `${parentId}/${first}_.${rest.join('.')}`
+        if (routeSegments.length > 1) {
+          routeSegments.shift()
+          routeId += `/${routeSegments.join('.')}`
+        }
+      } else {
+        routeId = `${parentId}/${routeSegments.join('.')}`
+      }
+    }
   }
+
+  if (convention === 'hybrid') {
+    let flat = routeId
+      // convert path separators /+ hybrid format
+      .replace(pathSepRegex, '+/')
+      // convert single _ to [_] due to conflict with new pathless layout prefix
+      .replace(/(^|\/|\.)_([^_.])/g, '$1[_]$2')
+      // convert double __ to single _ for pathless layout prefix
+      .replace(/(^|\/|\.)__/g, '$1_')
+      // convert index to _index for index routes
+      .replace(/(^|\/|\.)index$/, '$1_index')
+
+    // check if route is a parent route
+    // is so, move to hybrid folder (+) as _layout route
+    if (Object.values(routes).some(r => r.parentId === id)) {
+      flat = flat + '+/_layout'
+    }
+
+    return flat
+  }
+  // convert to flat route convention
+  let flat = routeId
+    // convert path separators to dots
+    .replace(pathSepRegex, '.')
+    // convert single _ to [_] due to conflict with new pathless layout prefix
+    .replace(/(^|\/|\.)_([^_.])/g, '$1[_]$2')
+    // convert double __ to single _ for pathless layout prefix
+    .replace(/(^|\/|\.)__/g, '$1_')
+    // convert index to _index for index routes
+    .replace(/(^|\/|\.)index$/, '$1_index')
+
+  return flat
 }
 
-export function convertToRoute(name: string) {
-  const pathSegments = name.split(path.sep)
-  return pathSegments
-    .map(pathSegment => {
-      const index = /(^|[.])index$/.test(pathSegment)
-      const routeSegments = getRouteSegments(pathSegment, index)
-      return getFlatRoute(routeSegments)
-    })
-    .join('.')
+function getRouteSegments(routePath: string) {
+  return routePath.split('/')
 }
 
-function getFlatRoute(segments: string[]) {
-  return segments
-    .map(segment =>
-      segment.startsWith('__')
-        ? segment.substring(1)
-        : segment === 'index'
-        ? '_index'
-        : segment,
-    )
-    .join('.')
+function getEscapedDottedRouteSegments(name: string) {
+  let routeSegments: string[] = []
+  let i = 0
+  let routeSegment = ''
+  let state = 'START'
+  let subState = 'NORMAL'
+
+  let pushRouteSegment = (routeSegment: string) => {
+    if (routeSegment) {
+      routeSegments.push(routeSegment)
+    }
+  }
+
+  while (i < name.length) {
+    let char = name[i]
+    switch (state) {
+      case 'START':
+        // process existing segment
+        pushRouteSegment(routeSegment)
+        routeSegment = ''
+        state = 'PATH'
+        continue // restart without advancing index
+      case 'PATH':
+        if (char === '.' && subState === 'NORMAL') {
+          state = 'START'
+          break
+        } else if (char === '[') {
+          subState = 'ESCAPE'
+          break
+        } else if (char === ']') {
+          subState = 'NORMAL'
+          break
+        }
+        routeSegment += char
+        break
+    }
+    i++ // advance to next character
+  }
+  // process remaining segment
+  pushRouteSegment(routeSegment)
+  return routeSegments
 }